refactor: rename ruleset folder to rulesets
SL-2295
