+++ src/cli/commands/lint.ts
@@ -3,9 +3,9 @@ import { IParserResult } from '@stoplight/types';
 import { getLocationForJsonPath } from '@stoplight/yaml';
 import { resolve } from 'path';
 import { readParsable } from '../../fs/reader';
-import { readRuleset } from '../../ruleset/reader';
 import { oas2Functions, oas2Rules } from '../../rulesets/oas2';
 import { oas3Functions, oas3Rules } from '../../rulesets/oas3';
+import { readRuleset } from '../../rulesets/reader';
 import { Spectral } from '../../spectral';
 import { IParsedResult, RuleCollection } from '../../types';
 import { formatOutput, writeOutput } from '../utils/output';
