+++ src/rulesets/__tests__/__fixtures__/valid-ruleset.json
